+++ CHANGES.rst
@@ -7,6 +7,8 @@ Release notes
 - Add Schematics 2.* support. Note that Schematics 2.0.0 introduced many
   changes to its API and even some validation rules have a slightly different
   behaviour in some cases.
+- ItemValidationPipeline optimisations for cases where no validators can be
+  applied.
 
 1.0.0 (2018-03-08)
 ------------------